#554: move all config to src/config
