+++ src/iohandler/curl_io_handler.cc
@@ -32,7 +32,7 @@
 #ifdef HAVE_CURL
 
 #include "curl_io_handler.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "util/tools.h"
 
 using namespace zmm;
