+++ src/layout/fallback_layout.cc
@@ -31,7 +31,7 @@
 
 #include "fallback_layout.h"
 #include "content_manager.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "metadata/metadata_handler.h"
 #include "util/string_converter.h"
 #include "util/tools.h"
