+++ src/metadata/libexif_handler.cc
@@ -33,7 +33,7 @@
 #ifdef HAVE_LIBEXIF
 
 #include "libexif_handler.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "iohandler/mem_io_handler.h"
 #include "util/tools.h"
