+++ src/metadata/taglib_handler.cc
@@ -47,7 +47,7 @@
 #include <taglib/vorbisfile.h>
 #include <taglib/wavpackfile.h>
 
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "iohandler/mem_io_handler.h"
 #include "util/string_converter.h"
 #include "taglib_handler.h"
