+++ src/onlineservice/online_service_helper.cc
@@ -33,7 +33,7 @@
 #ifdef ONLINE_SERVICES
 
 #include "online_service_helper.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "mxml/mxml.h"
 #include "online_service.h"
 #include "zmm/zmm.h"
