+++ src/upnp_cds.cc
@@ -30,7 +30,7 @@
 /// \file upnp_cds.cc
 
 #include "upnp_cds.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "search_handler.h"
 #include "server.h"
 #include "storage.h"
