+++ src/url.cc
@@ -33,7 +33,7 @@
 
 #include "url.h"
 #include "common.h"
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "util/rexp.h"
 #include "util/tools.h"
 #include <pthread.h>
