+++ src/util/tools.cc
@@ -56,7 +56,7 @@
 #include <sys/sockio.h>
 #endif
 
-#include "config_manager.h"
+#include "config/config_manager.h"
 #include "contrib/md5.h"
 #include "iohandler/file_io_handler.h"
 #include "metadata/metadata_handler.h"
